refactor(referral): extract HardwareRecordsList and fix timeline border
Extract HardwareRecordsList into a separate component file for better
code organization. Update timeline section to use neutral/2 border
color on both top and bottom edges per design spec.
