+++ packages/kit/src/views/ReferFriends/pages/HardwareSalesReward/index.tsx
@@ -4,7 +4,6 @@ import { useIntl } from 'react-intl';
 
 import {
   Alert,
-  Empty,
   Page,
   RefreshControl,
   ScrollView,
@@ -30,61 +29,8 @@ import { EAccountSelectorSceneName } from '@onekeyhq/shared/types';
 
 import { BreadcrumbSection, ReferFriendsPageContainer } from '../../components';
 
-import { HardwareRecordCard } from './components/HardwareRecordCard';
-import { HardwareRecordTable } from './components/HardwareRecordTable';
+import { HardwareRecordsList } from './components/HardwareRecordsList';
 import { HardwareSalesRewardHeader } from './components/HardwareSalesRewardHeader';
-
-function HardwareRecordsList({
-  isLoading,
-  records,
-  isMobile,
-}: {
-  isLoading: boolean;
-  records: IHardwareRecordItem[];
-  isMobile: boolean;
-}) {
-  const intl = useIntl();
-
-  if (isLoading && records.length === 0) {
-    return (
-      <YStack ai="center" jc="center" py="$10" px="$5">
-        <Spinner size="large" />
-      </YStack>
-    );
-  }
-
-  if (records.length === 0) {
-    return (
-      <YStack px="$5" gap="$3" pb="$5">
-        <Empty
-          icon="GiftOutline"
-          title={intl.formatMessage({
-            id: ETranslations.referral_referred_empty,
-          })}
-          description={intl.formatMessage({
-            id: ETranslations.referral_referred_empty_desc,
-          })}
-        />
-      </YStack>
-    );
-  }
-
-  if (isMobile) {
-    return (
-      <YStack px="$5" gap="$3" pb="$5">
-        {records.map((record) => (
-          <HardwareRecordCard key={record._id} item={record} />
-        ))}
-      </YStack>
-    );
-  }
-
-  return (
-    <YStack px="$5" gap="$3" pb="$5">
-      <HardwareRecordTable records={records} />
-    </YStack>
-  );
-}
 
 function HardwareSalesRewardPageWrapper() {
   // Redirect to ReferAFriend page if user is not logged in
